Change id type, int/string to long

+++ FlySMM/src/main/java/promotion/Promotion.java
@@ -3,7 +3,7 @@ package promotion;
 public abstract class Promotion {
 	protected int discountRate;
 	protected boolean fidelity;
-	protected int idPromo;
+	protected long idPromo;
 	protected String name;
 	
 	public abstract String notify_();
